websocket_sender/requirements.txt
aio-pika==9.4.1
backoff==2.2.1
pydantic==2.6.4
pydantic-settings==2.2.1
PyJWT==2.8.0
websockets==12.0.0

faker==24.4.0
